move CreateDefaultConfigurationSource() from Tracer to TracerConfiguration
